work in progress (expr)

+++ expr-lib/src/main/java/fr/an/exprlib/sql/analysis/ExprDTOToSimpleExprConverter.java
@@ -33,7 +33,7 @@ public class ExprDTOToSimpleExprConverter {
             if (expr.intValue != null) {
                 return new SimpleExpr.LiteralExpr(expr.intValue);
             }else if (expr.doubleValue != null) {
-                return new SimpleExpr.LiteralExpr(expr.intValue);
+                return new SimpleExpr.LiteralExpr(expr.doubleValue);
             } else if (expr.strValue != null) {
                 return new SimpleExpr.LiteralExpr(expr.strValue);
             } else {
